ODYS-802 use generics and remove unneeded params

+++ src/org/tinyradius/attribute/IntegerAttribute.java
@@ -82,7 +82,7 @@ public class IntegerAttribute extends RadiusAttribute {
 		if (at != null) {
 			Integer val = at.getEnumeration(value);
 			if (val != null) {
-				setAttributeValue(val.intValue());
+				setAttributeValue(val);
 				return;
 			}
 		}
